Working camera + Frame work for 2 games
When you now take a picture you can now view the whole screen before taking a picture, and when you go to view the image, it will be displayed correctly without anything getting in the way, still want to make save the to gallery and delete function if possible.

In addition, I have created the framework for both games on the app first will be a frit ninja clone and the second will be a doodle jump clone, but so far for the second game this has been troubling and annoying, I might need to rethink both games.