Working framework for game 1 + Serious error found?
I was able to create the functions and layout for the first game called "Negative thought slicer" based on Fruit Ninja.

However I noticed that for an unknown reason, if I duplicated the scene from Main menu and repurposed it for anything I wanted, key features for game play do not work at all.

But If I create a new scene and copy all of my game elements into it and use it, it works just fine.

I have been unable to figure out the cause of the problem but only know this solution for the time being.