Added UI to game 1 + need to fix score error
I've taken the time to add in a timer, score, and other UI features to game 1.

However the score counter does not work correctly for an unknown reason.

Every time you slice negative thoughts with your blade, 1 point is added and shown on the score UI text.

But an error occurs making this impossible, whilst the rest of the game still works just as fine.