Fixed Score error + other errors
I fixed the error with the score, the problem was that I did not know that I was incorrectly trying to reference a function that would increase the score int and display it on screen.

Instead of referencing it the normal way with in the script by declaring the name of the variable from the other script, I needed to use "FindObjectOfType", which would actively tell the script where to find the function within the other script every time the blade came into contact with the sliceable object.

In-addition I had to fix a few minor UI bugs to allow the game to preform a full game play loop without causing a soft lock of states through overlapping bools.