More improvements to the chat bot
I've made the chat bot and it is able to talk to the user by recognising keywords typed out, and producing an automated response.

I've also included shortcuts to parts of the app the user would be interested in seeing.

However more improvements could be made to the systems  automated response for detecting keywords